LineTraceReflect
Using the C++ version of the default First-Person shooter Unreal template, you are to implement a laser, fired from the player, that reflects off of what it hits. Create a variable called NumberOfBounces and ensure that the value can be changed in the editor. When the laser bounces off a wall, have it continue to bounce off of subsequent walls a number of times equal to NumberOfBounces